refactor: Modularize hardhat config by moving tasks to dedicated task files
feat: Update protocol improvement task to PI-13
